requirements.txt
streamlit
google-generativeai
numpy